docs: update refactoring progress after citation_verify completion
Update refactoring_opportunities.md to reflect the successful completion
of the citation_verify.py refactoring work completed on 2025-10-23.

Key updates:
- Mark citation_verify.py refactoring as completed (914 lines → 8 focused modules)
- Update executive summary: 2 high-priority files now refactored (down from 1)
- Reduce remaining large files count from 3 to 2
- Document new citation/ package structure with 8 modules
- Update next priority to commands/verify.py (829 lines)
- Note deep coupling chain has been successfully untangled
- Confirm all 387 unit tests still passing with zero breaking changes
